iot/requirements.txt
paho-mqtt
pvporcupine
pyaudio
requests
numpy
aiohttp
grpcio-tools